+++ src/main/java/com/hydata/intelligence/platform/controller/DataStreamModelController.java
@@ -23,6 +23,7 @@ import com.hydata.intelligence.platform.model.DataStreamModel;
 import com.hydata.intelligence.platform.model.RESCODE;
 import com.hydata.intelligence.platform.repositories.UnitTypeRepository;
 import com.hydata.intelligence.platform.service.DataStreamModelService;
+import com.hydata.intelligence.platform.utils.CheckParams;
 
 
 /**
@@ -47,8 +48,16 @@ public class DataStreamModelController {
 		}
 	 
 	 @RequestMapping(value = "/delete" ,method = RequestMethod.GET)
-	    public JSONObject delete(int id){	    		
-	    	return dsmService.deleteByDSMId(id);
+	    public JSONObject delete(int id){
+		 JSONObject params = new JSONObject();
+			params.put("id", id);
+			JSONObject result = CheckParams.checkParams(params);
+			if((Integer)result.get("code")==0) {			
+				return dsmService.deleteByDSMId(id);
+			}else {
+				return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+			}
+	    	
 	    }
 	 @RequestMapping(value = "/modify" ,method = RequestMethod.POST)
 	    public JSONObject modify(@RequestBody DataStreamModel model){	    	
@@ -62,35 +71,64 @@ public class DataStreamModelController {
 	    }
 	 @RequestMapping(value = "/SimPage",method = RequestMethod.GET)
 	 	public JSONObject pageable(Long productId,Integer page,Integer number) {
-		 
-		 Page<DatastreamModel> DatastreamModelPage = dsmService.queryByProductId(productId,page-1,number);
-         List<DatastreamModel> DatastreamModelList = DatastreamModelPage.getContent();
-         List<DataStreamModel> Data_stream_modelList = new ArrayList<>();
-         for(DatastreamModel model:DatastreamModelList) {
-        	 DataStreamModel mod = new DataStreamModel();
-        	 mod.setId(model.getId());
-        	 mod.setProduct_id(model.getProductId());
-        	 mod.setName(model.getName());
-        	 mod.setCreateTime(model.getCreateTime());
-        	 Optional<UnitType> unitOptional = unitTypeRepository.findById(model.getUnitTypeId());
-        	 if(unitOptional.isPresent()) {
-        		 UnitType unit  = unitOptional.get();
-        		 mod.setUnit_name(unit.getName());
-        		 mod.setUnit_symbol(unit.getSymbol());
-        	 }
-        	 Data_stream_modelList.add(mod);
-         }
-         return RESCODE.SUCCESS.getJSONRES(Data_stream_modelList, DatastreamModelPage.getTotalPages(), DatastreamModelPage.getTotalElements());
+		 JSONObject params = new JSONObject();
+			params.put("productId", productId);
+			params.put("page", page);
+			params.put("number", number);
+			JSONObject result = CheckParams.checkParams(params);
+			if((Integer)result.get("code")==0) {			
+				
+				Page<DatastreamModel> DatastreamModelPage = dsmService.queryByProductId(productId,page-1,number);
+		         List<DatastreamModel> DatastreamModelList = DatastreamModelPage.getContent();
+		         List<DataStreamModel> Data_stream_modelList = new ArrayList<>();
+		         for(DatastreamModel model:DatastreamModelList) {
+		        	 DataStreamModel mod = new DataStreamModel();
+		        	 mod.setId(model.getId());
+		        	 mod.setProduct_id(model.getProductId());
+		        	 mod.setName(model.getName());
+		        	 mod.setCreateTime(model.getCreateTime());
+		        	 Optional<UnitType> unitOptional = unitTypeRepository.findById(model.getUnitTypeId());
+		        	 if(unitOptional.isPresent()) {
+		        		 UnitType unit  = unitOptional.get();
+		        		 mod.setUnit_name(unit.getName());
+		        		 mod.setUnit_symbol(unit.getSymbol());
+		        	 }
+		        	 Data_stream_modelList.add(mod);
+		         }
+		         return RESCODE.SUCCESS.getJSONRES(Data_stream_modelList, DatastreamModelPage.getTotalPages(), DatastreamModelPage.getTotalElements());
+				
+			}else {
+				return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+			}
 	 }
 	 @RequestMapping(value ="/find_by_name",method = RequestMethod.GET)
 	 public JSONObject findByName(Integer page,Integer number,String dsmName,Long productId) {
-		 Page<DatastreamModel> DatastreamModelPage = dsmService.findByName(page-1,number,dsmName,productId);
-		 return RESCODE.SUCCESS.getJSONRES(DatastreamModelPage.getContent(),DatastreamModelPage.getTotalPages(),DatastreamModelPage.getTotalElements());
+		JSONObject params = new JSONObject();
+		params.put("page", page);
+		params.put("number", number);
+		params.put("dsmName", dsmName);
+		params.put("productId", productId);
+		JSONObject result = CheckParams.checkParams(params);
+		if((Integer)result.get("code")==0) {			
+			 Page<DatastreamModel> DatastreamModelPage = dsmService.findByName(page-1,number,dsmName,productId);
+			 return RESCODE.SUCCESS.getJSONRES(DatastreamModelPage.getContent(),DatastreamModelPage.getTotalPages(),DatastreamModelPage.getTotalElements());
+		}else {
+			return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+		}		
 	 }
 	 
 	 @RequestMapping(value ="/get_increment",method = RequestMethod.GET)
 	 public JSONObject getIncrement(Long product_id,String start,String end) throws ParseException {		 
-	     return dsmService.getIncrement(product_id, sdf.parse(start), sdf.parse(end));
+		 	JSONObject params = new JSONObject();
+			params.put("product_id", product_id);
+			params.put("start", start);
+			params.put("end", end);
+			JSONObject result = CheckParams.checkParams(params);
+			if((Integer)result.get("code")==0) {			
+				return dsmService.getIncrement(product_id, sdf.parse(start), sdf.parse(end));
+			}else {
+				return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+			}	     
 	 }
 
 }
