+++ src/main/java/com/hydata/intelligence/platform/controller/DeviceController.java
@@ -12,7 +12,9 @@ import org.springframework.web.bind.annotation.*;
 
 import com.alibaba.fastjson.JSONObject;
 import com.hydata.intelligence.platform.dto.Device;
+import com.hydata.intelligence.platform.model.RESCODE;
 import com.hydata.intelligence.platform.service.DeviceService;
+import com.hydata.intelligence.platform.utils.CheckParams;
 import com.hydata.intelligence.platform.utils.ExcelUtils;
 
 import javax.servlet.http.HttpServletRequest;
@@ -36,7 +38,17 @@ public class DeviceController {
 	
 	@RequestMapping(value="/show",method=RequestMethod.GET)
 	public JSONObject showAll(Long product_id,Integer page,Integer number,int sort){
-		return deviceService.showAllByProductIdM(product_id, page, number,sort);
+		JSONObject params = new JSONObject();
+		params.put("product_id", product_id);
+		params.put("page", page);
+		params.put("number", number);
+		params.put("sort", sort);
+		JSONObject result = CheckParams.checkParams(params);
+		if((Integer)result.get("code")==0) {			
+			return deviceService.showAllByProductIdM(product_id, page, number,sort);
+		}else {
+			return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+		}			
 	}
 	
 	@RequestMapping(value ="/add",method=RequestMethod.POST)
@@ -46,7 +58,17 @@ public class DeviceController {
 	
 	@RequestMapping(value = "/query_by_sn_or_name",method = RequestMethod.GET)
 	public JSONObject queryDeviceByDevice_snOrName(Long product_id,Integer page,Integer number,String device_snOrName){
-		return deviceService.queryByDeviceSnOrName_m(product_id,device_snOrName, page, number);
+		JSONObject params = new JSONObject();
+		params.put("product_id", product_id);
+		params.put("page", page);
+		params.put("number", number);
+		params.put("device_snOrName", device_snOrName);
+		JSONObject result = CheckParams.checkParams(params);
+		if((Integer)result.get("code")==0) {			
+			return deviceService.queryByDeviceSnOrName_m(product_id,device_snOrName, page, number);
+		}else {
+			return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+		}		
 	}
 	
 	@RequestMapping(value="/modify",method=RequestMethod.POST)
@@ -56,17 +78,40 @@ public class DeviceController {
 	
 	@RequestMapping(value="/delete",method=RequestMethod.GET)
 	public JSONObject delete(String device_sn){
-		return deviceService.deleteDevice(device_sn);
+		JSONObject params = new JSONObject();
+		params.put("device_sn", device_sn);
+		JSONObject result = CheckParams.checkParams(params);
+		if((Integer)result.get("code")==0) {			
+			return deviceService.deleteDevice(device_sn);
+		}else {
+			return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+		}		
 	}
 	
 	@RequestMapping(value="/get_devicelist",method = RequestMethod.GET)
 	public JSONObject getByProductId(Integer productId) {
-		return deviceService.getByProductId(productId);
+		JSONObject params = new JSONObject();
+		params.put("productId", productId);
+		JSONObject result = CheckParams.checkParams(params);
+		if((Integer)result.get("code")==0) {			
+			return deviceService.getByProductId(productId);
+		}else {
+			return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+		}
+		
 	}
 	
 	@RequestMapping(value="/get_devicedslist",method = RequestMethod.GET)
 	public JSONObject getDeviceDsByDeviceSn(String device_sn) {
-		return deviceService.getDeviceDsByDeviceSn(device_sn);
+		JSONObject params = new JSONObject();
+		params.put("device_sn", device_sn);
+		JSONObject result = CheckParams.checkParams(params);
+		if((Integer)result.get("code")==0) {			
+			return deviceService.getDeviceDsByDeviceSn(device_sn);
+		}else {
+			return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+		}
+		
 	}
 
 	@RequestMapping(value = "/export_excel",method=RequestMethod.GET)
@@ -76,23 +121,59 @@ public class DeviceController {
 	
 	@RequestMapping(value = "/import_excel",method=RequestMethod.GET)
 	public JSONObject importExcel(String url,Long productId) {
-		return deviceService.importExcel(url, productId);	
+		JSONObject params = new JSONObject();
+		params.put("url", url);
+		params.put("productId", productId);
+		JSONObject result = CheckParams.checkParams(params);
+		if((Integer)result.get("code")==0) {			
+			return deviceService.importExcel(url, productId);	
+		}else {
+			return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+		}
+		
 	}
 	
 	@RequestMapping(value= "/get_increment",method = RequestMethod.GET)
 	public JSONObject getIncrement(Long product_id,String start ,String end) throws ParseException {
-		SimpleDateFormat sdf = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss");
-		return deviceService.getIncrement(product_id, sdf.parse(start), sdf.parse(end));
+		JSONObject params = new JSONObject();
+		params.put("product_id", product_id);
+		params.put("start", start);
+		params.put("end", end);
+		JSONObject result = CheckParams.checkParams(params);
+		if((Integer)result.get("code")==0) {			
+			SimpleDateFormat sdf = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss");
+			return deviceService.getIncrement(product_id, sdf.parse(start), sdf.parse(end));
+		}else {
+			return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+		}		
 	}
 	
 	@RequestMapping(value= "/get_cmd_logs",method = RequestMethod.GET)
 	public JSONObject getCmdLogs(String device_sn) {
-		return deviceService.getCmdLogs(device_sn);
+		JSONObject params = new JSONObject();
+		params.put("device_sn", device_sn);
+		JSONObject result = CheckParams.checkParams(params);
+		if((Integer)result.get("code")==0) {			
+			return deviceService.getCmdLogs(device_sn);
+		}else {
+			return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+		}
+		
 	}
 	
 	@RequestMapping(value= "/get_device_ds_data",method = RequestMethod.GET)
 	public JSONObject getDeviceDsData(Integer dd_id,Date start,Date end) {
-		return deviceService.getDeviceDsData(dd_id, start, end);
+		JSONObject params = new JSONObject();
+		params.put("dd_id", dd_id);
+		params.put("start", start);
+		params.put("end", end);
+		JSONObject result = CheckParams.checkParams(params);
+		if((Integer)result.get("code")==0) {			
+			return deviceService.getDeviceDsData(dd_id, start, end);
+		}else {
+			return RESCODE.PARAM_MISSING.getJSONRES(result.get("data"));
+		}
+		
 	}
 	
 	@RequestMapping(value= "/get_data")
